+++ tree/testUtil.go
@@ -5,7 +5,7 @@ package tree
 // in a tree in order of selected traversal.
 
 // returns the primary keys of a tree in order of a breadth first search
-func bfc(q []Node, iter []uint64) []uint64 {
+func bfc(q []Node, iter []uint) []uint {
 	iter = append(iter, q[0].GetID())
 	q = append(q[1:], q[0].GetChildren()...)
 	if len(q) == 0 {
@@ -15,7 +15,7 @@ func bfc(q []Node, iter []uint64) []uint64 {
 }
 
 // returns the primary keys of a tree in order of a breadth first search
-func dfc(n Node, iter []uint64) []uint64 {
+func dfc(n Node, iter []uint) []uint {
 	iter = append(iter, n.GetID())
 	for _, c := range n.GetChildren() {
 		iter = dfc(c, iter)
